requirements.txt
discord.py
flask
g4f
typing_extensions
revampdb
